Refactor dataset revision endpoint to use new DatasetRevisionService and update schemas
