+++ application/backend/app/api/routers/dataset_revisions.py
@@ -8,13 +8,13 @@ from uuid import UUID
 from fastapi import APIRouter, Depends, Query, status
 from starlette.responses import FileResponse
 
-from app.api.dependencies import get_dataset_revision, get_dataset_service, get_project
-from app.api.schemas.dataset_item import DatasetItemsWithPagination, DatasetItemView
+from app.api.dependencies import get_dataset_revision, get_dataset_revision_service, get_project
+from app.api.schemas.dataset_item import DatasetItemRevisionView, DatasetItemsWithPagination
 from app.api.validators import DatasetItemID, DatasetRevisionID
 from app.core.models import Pagination
 from app.models import DatasetItemFormat, DatasetItemSubset, Project
 from app.models.dataset_revision import DatasetRevision
-from app.services import DatasetService
+from app.services import DatasetRevisionService
 
 router = APIRouter(
     prefix="/api/projects/{project_id}/dataset_revisions/{dataset_revision_id}",
@@ -37,14 +37,14 @@ MAX_DATASET_ITEMS_NUMBER_RETURNED = 100
 )
 def list_dataset_revision_items(
     project: Annotated[Project, Depends(get_project)],
-    dataset_service: Annotated[DatasetService, Depends(get_dataset_service)],
+    dataset_revision_service: Annotated[DatasetRevisionService, Depends(get_dataset_revision_service)],
     dataset_revision: Annotated[DatasetRevision, Depends(get_dataset_revision)],
     limit: Annotated[int, Query(ge=1, le=MAX_DATASET_ITEMS_NUMBER_RETURNED)] = DEFAULT_DATASET_ITEMS_NUMBER_RETURNED,
     offset: Annotated[int, Query(ge=0)] = 0,
     subset: Annotated[DatasetItemSubset | None, Query()] = None,
 ) -> DatasetItemsWithPagination:
     """List the items in a dataset revision. This endpoint supports pagination."""
-    items_data, total_count = dataset_service.list_dataset_revision_items(
+    items_data, total_count = dataset_revision_service.list_dataset_revision_items(
         project_id=project.id,
         revision_id=dataset_revision.id,
         limit=limit,
@@ -54,14 +54,12 @@ def list_dataset_revision_items(
 
     items = []
     for item_data in items_data:
-        item_view = DatasetItemView(
+        item_view = DatasetItemRevisionView(
             id=UUID(item_data["id"]),
             name=Path(item_data["image"]).stem,
             format=DatasetItemFormat(Path(item_data["image"]).suffix.lstrip(".")),
             width=item_data["image_info"]["width"],
             height=item_data["image_info"]["height"],
-            size=0,  # Not available in parquet
-            source_id=None,
             subset=DatasetItemSubset(item_data["subset"]),
         )
         items.append(item_view)
@@ -80,7 +78,7 @@ def list_dataset_revision_items(
 @router.get(
     "/items/{dataset_item_id}",
     responses={
-        status.HTTP_200_OK: {"description": "Dataset item found", "model": DatasetItemView},
+        status.HTTP_200_OK: {"description": "Dataset item found", "model": DatasetItemRevisionView},
         status.HTTP_400_BAD_REQUEST: {"description": "Invalid dataset item ID or revision ID"},
         status.HTTP_404_NOT_FOUND: {"description": "Dataset item, revision, or project not found"},
     },
@@ -89,23 +87,21 @@ def get_dataset_revision_item(
     project: Annotated[Project, Depends(get_project)],
     dataset_revision: Annotated[DatasetRevision, Depends(get_dataset_revision)],
     dataset_item_id: DatasetItemID,
-    dataset_service: Annotated[DatasetService, Depends(get_dataset_service)],
-) -> DatasetItemView:
+    dataset_revision_service: Annotated[DatasetRevisionService, Depends(get_dataset_revision_service)],
+) -> DatasetItemRevisionView:
     """Get information about a specific item in the dataset revision"""
-    item_data = dataset_service.get_dataset_revision_item(
+    item_data = dataset_revision_service.get_dataset_revision_item(
         project_id=project.id,
         revision_id=dataset_revision.id,
         item_id=str(dataset_item_id),
     )
 
-    return DatasetItemView(
+    return DatasetItemRevisionView(
         id=UUID(item_data["id"]),
         name=Path(item_data["image"]).stem,
         format=DatasetItemFormat(Path(item_data["image"]).suffix.lstrip(".")),
         width=item_data["image_info"]["width"],
         height=item_data["image_info"]["height"],
-        size=0,  # Not available in parquet
-        source_id=None,
         subset=DatasetItemSubset(item_data["subset"]),
     )
 
@@ -122,10 +118,10 @@ def get_dataset_revision_item_binary(
     project: Annotated[Project, Depends(get_project)],
     dataset_revision: Annotated[DatasetRevision, Depends(get_dataset_revision)],
     dataset_item_id: DatasetItemID,
-    dataset_service: Annotated[DatasetService, Depends(get_dataset_service)],
+    dataset_revision_service: Annotated[DatasetRevisionService, Depends(get_dataset_revision_service)],
 ) -> FileResponse:
     """Get the image data of an item in the dataset revision"""
-    binary_path = dataset_service.get_dataset_revision_item_binary_path(
+    binary_path = dataset_revision_service.get_dataset_revision_item_binary_path(
         project_id=project.id,
         revision_id=dataset_revision.id,
         item_id=str(dataset_item_id),
@@ -145,10 +141,10 @@ def get_dataset_revision_item_thumbnail(
     project: Annotated[Project, Depends(get_project)],
     dataset_revision: Annotated[DatasetRevision, Depends(get_dataset_revision)],
     dataset_item_id: DatasetItemID,
-    dataset_service: Annotated[DatasetService, Depends(get_dataset_service)],
+    dataset_revision_service: Annotated[DatasetRevisionService, Depends(get_dataset_revision_service)],
 ) -> FileResponse:
     """Get the thumbnail of an item in the dataset revision"""
-    binary_path = dataset_service.get_dataset_revision_item_binary_path(
+    binary_path = dataset_revision_service.get_dataset_revision_item_binary_path(
         project_id=project.id,
         revision_id=dataset_revision.id,
         item_id=str(dataset_item_id),
@@ -169,7 +165,7 @@ def delete_dataset_revision_files(
     project: Annotated[Project, Depends(get_project)],
     dataset_revision_id: DatasetRevisionID,
     _dataset_revision: Annotated[DatasetRevision, Depends(get_dataset_revision)],
-    dataset_service: Annotated[DatasetService, Depends(get_dataset_service)],
+    dataset_revision_service: Annotated[DatasetRevisionService, Depends(get_dataset_revision_service)],
 ) -> None:
     """Delete the files associated with a dataset revision"""
-    dataset_service.delete_dataset_revision_files(project_id=project.id, revision_id=dataset_revision_id)
+    dataset_revision_service.delete_dataset_revision_files(project_id=project.id, revision_id=dataset_revision_id)
