+++ application/backend/app/api/schemas/dataset_item.py
@@ -10,6 +10,30 @@ from app.core.models import BaseRequiredIDNameModel, Pagination
 from app.models import DatasetItemAnnotation, DatasetItemFormat, DatasetItemSubset
 
 
+class DatasetItemRevisionView(BaseRequiredIDNameModel):
+    """
+    Dataset Revision item
+    """
+
+    format: DatasetItemFormat
+    width: int
+    height: int
+    subset: DatasetItemSubset
+
+    model_config = {
+        "json_schema_extra": {
+            "example": {
+                "id": "7b073838-99d3-42ff-9018-4e901eb047fc",
+                "name": "img-010203",
+                "format": "jpg",
+                "width": 1280,
+                "height": 720,
+                "subset": "unassigned",
+            }
+        }
+    }
+
+
 class DatasetItemView(BaseRequiredIDNameModel):
     """
     Dataset item
@@ -87,7 +111,7 @@ class DatasetItemsWithPagination(BaseModel):
     Dataset Items list with pagination info
     """
 
-    items: list[DatasetItemView]
+    items: list[DatasetItemView | DatasetItemRevisionView]
     pagination: Pagination
