+++ application/backend/app/lifecycle.py
@@ -19,7 +19,13 @@ from app.core.run import Runnable, RunnableFactory
 from app.db import MigrationManager, get_db_session
 from app.scheduler import Scheduler
 from app.schemas.job import JobType
-from app.services import DatasetService, LabelService, ModelService, TrainingConfigurationService
+from app.services import (
+    DatasetRevisionService,
+    DatasetService,
+    LabelService,
+    ModelService,
+    TrainingConfigurationService,
+)
 from app.services.base_weights_service import BaseWeightsService
 from app.services.data_collect import DataCollector
 from app.services.event.event_bus import EventBus
@@ -56,6 +62,7 @@ def setup_job_controller(data_dir: Path, max_parallel_jobs: int) -> tuple[JobQue
                 subset_service=SubsetService(),
                 subset_assigner=SubsetAssigner(),
                 dataset_service=DatasetService(data_dir=data_dir, label_service=LabelService()),
+                dataset_revision_service=DatasetRevisionService(data_dir=data_dir),
                 model_service=ModelService(data_dir=data_dir),
                 training_configuration_service=TrainingConfigurationService(),
                 data_dir=data_dir,
