+++ application/backend/tests/unit/routers/test_dataset_revisions.py
@@ -9,11 +9,11 @@ from uuid import uuid4
 import pytest
 from fastapi import status
 
-from app.api.dependencies import get_dataset_service
+from app.api.dependencies import get_dataset_revision_service
 from app.api.schemas.dataset_item import DatasetItemSubset
 from app.main import app
 from app.models import DatasetItem, DatasetItemFormat
-from app.services import DatasetService, ResourceNotFoundError, ResourceType
+from app.services import DatasetRevisionService, ResourceNotFoundError, ResourceType
 
 
 @pytest.fixture
@@ -41,17 +41,17 @@ def fxt_dataset_item():
 
 
 @pytest.fixture
-def fxt_dataset_service() -> MagicMock:
-    dataset_service = MagicMock(spec=DatasetService)
-    app.dependency_overrides[get_dataset_service] = lambda: dataset_service
-    return dataset_service
+def fxt_dataset_revision_service() -> MagicMock:
+    dataset_revision_service = MagicMock(spec=DatasetRevisionService)
+    app.dependency_overrides[get_dataset_revision_service] = lambda: dataset_revision_service
+    return dataset_revision_service
 
 
 class TestDatasetRevisionItemEndpoints:
     def test_list_dataset_revision_items_revision_not_found(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_service, fxt_client
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_revision_service, fxt_client
     ):
-        fxt_dataset_service.get_dataset_revision.side_effect = ResourceNotFoundError(
+        fxt_dataset_revision_service.get_dataset_revision.side_effect = ResourceNotFoundError(
             ResourceType.DATASET_REVISION, str(fxt_dataset_revision_id)
         )
 
@@ -60,22 +60,22 @@ class TestDatasetRevisionItemEndpoints:
         )
 
         assert response.status_code == status.HTTP_404_NOT_FOUND
-        fxt_dataset_service.get_dataset_revision.assert_called_once_with(
+        fxt_dataset_revision_service.get_dataset_revision.assert_called_once_with(
             project_id=fxt_get_project.id, revision_id=fxt_dataset_revision_id
         )
-        fxt_dataset_service.list_dataset_revision_items.assert_not_called()
+        fxt_dataset_revision_service.list_dataset_revision_items.assert_not_called()
 
     def test_list_dataset_revision_items_success(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_item, fxt_dataset_service, fxt_client
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_item, fxt_dataset_revision_service, fxt_client
     ):
-        fxt_dataset_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
+        fxt_dataset_revision_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
         mock_item_data = {
             "id": str(fxt_dataset_item.id),
             "image": f"/path/to/{fxt_dataset_item.name}.jpg",
             "image_info": {"width": fxt_dataset_item.width, "height": fxt_dataset_item.height},
             "subset": fxt_dataset_item.subset.value,
         }
-        fxt_dataset_service.list_dataset_revision_items.return_value = ([mock_item_data], 1)
+        fxt_dataset_revision_service.list_dataset_revision_items.return_value = ([mock_item_data], 1)
 
         response = fxt_client.get(
             f"/api/projects/{str(fxt_get_project.id)}/dataset_revisions/{str(fxt_dataset_revision_id)}/items"
@@ -91,10 +91,10 @@ class TestDatasetRevisionItemEndpoints:
         assert response_data["items"][0]["id"] == str(fxt_dataset_item.id)
         assert response_data["items"][0]["name"] == "test_dataset_item"
 
-        fxt_dataset_service.get_dataset_revision.assert_called_once_with(
+        fxt_dataset_revision_service.get_dataset_revision.assert_called_once_with(
             project_id=fxt_get_project.id, revision_id=fxt_dataset_revision_id
         )
-        fxt_dataset_service.list_dataset_revision_items.assert_called_once_with(
+        fxt_dataset_revision_service.list_dataset_revision_items.assert_called_once_with(
             project_id=fxt_get_project.id,
             revision_id=fxt_dataset_revision_id,
             limit=10,
@@ -103,16 +103,16 @@ class TestDatasetRevisionItemEndpoints:
         )
 
     def test_list_dataset_revision_items_with_pagination(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_item, fxt_dataset_service, fxt_client
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_item, fxt_dataset_revision_service, fxt_client
     ):
-        fxt_dataset_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
+        fxt_dataset_revision_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
         mock_item_data = {
             "id": str(fxt_dataset_item.id),
             "image": f"/path/to/{fxt_dataset_item.name}.jpg",
             "image_info": {"width": fxt_dataset_item.width, "height": fxt_dataset_item.height},
             "subset": fxt_dataset_item.subset.value,
         }
-        fxt_dataset_service.list_dataset_revision_items.return_value = ([mock_item_data], 100)
+        fxt_dataset_revision_service.list_dataset_revision_items.return_value = ([mock_item_data], 100)
 
         response = fxt_client.get(
             f"/api/projects/{str(fxt_get_project.id)}/dataset_revisions/{str(fxt_dataset_revision_id)}/items?limit=50&offset=10"
@@ -124,7 +124,7 @@ class TestDatasetRevisionItemEndpoints:
         assert response_data["pagination"]["limit"] == 50
         assert response_data["pagination"]["offset"] == 10
 
-        fxt_dataset_service.list_dataset_revision_items.assert_called_once_with(
+        fxt_dataset_revision_service.list_dataset_revision_items.assert_called_once_with(
             project_id=fxt_get_project.id,
             revision_id=fxt_dataset_revision_id,
             limit=50,
@@ -134,23 +134,29 @@ class TestDatasetRevisionItemEndpoints:
 
     @pytest.mark.parametrize("subset", ["unassigned", "training", "validation", "testing"])
     def test_list_dataset_revision_items_with_subset_filter(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_item, fxt_dataset_service, fxt_client, subset
+        self,
+        fxt_get_project,
+        fxt_dataset_revision_id,
+        fxt_dataset_item,
+        fxt_dataset_revision_service,
+        fxt_client,
+        subset,
     ):
-        fxt_dataset_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
+        fxt_dataset_revision_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
         mock_item_data = {
             "id": str(fxt_dataset_item.id),
             "image": f"/path/to/{fxt_dataset_item.name}.jpg",
             "image_info": {"width": fxt_dataset_item.width, "height": fxt_dataset_item.height},
             "subset": subset,
         }
-        fxt_dataset_service.list_dataset_revision_items.return_value = ([mock_item_data], 1)
+        fxt_dataset_revision_service.list_dataset_revision_items.return_value = ([mock_item_data], 1)
 
         response = fxt_client.get(
             f"/api/projects/{str(fxt_get_project.id)}/dataset_revisions/{str(fxt_dataset_revision_id)}/items?subset={subset}"
         )
 
         assert response.status_code == status.HTTP_200_OK
-        fxt_dataset_service.list_dataset_revision_items.assert_called_once_with(
+        fxt_dataset_revision_service.list_dataset_revision_items.assert_called_once_with(
             project_id=fxt_get_project.id,
             revision_id=fxt_dataset_revision_id,
             limit=10,
@@ -160,43 +166,45 @@ class TestDatasetRevisionItemEndpoints:
 
     @pytest.mark.parametrize("limit", [1000, 0, -20])
     def test_list_dataset_revision_items_invalid_limit(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_service, fxt_client, limit
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_revision_service, fxt_client, limit
     ):
         response = fxt_client.get(
             f"/api/projects/{str(fxt_get_project.id)}/dataset_revisions/{str(fxt_dataset_revision_id)}/items?limit={limit}"
         )
 
         assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
-        fxt_dataset_service.list_dataset_revision_items.assert_not_called()
+        fxt_dataset_revision_service.list_dataset_revision_items.assert_not_called()
 
     @pytest.mark.parametrize("offset", [-1, -20])
     def test_list_dataset_revision_items_invalid_offset(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_service, fxt_client, offset
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_revision_service, fxt_client, offset
     ):
         response = fxt_client.get(
             f"/api/projects/{str(fxt_get_project.id)}/dataset_revisions/{str(fxt_dataset_revision_id)}/items?offset={offset}"
         )
 
         assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
-        fxt_dataset_service.list_dataset_revision_items.assert_not_called()
+        fxt_dataset_revision_service.list_dataset_revision_items.assert_not_called()
 
-    def test_list_dataset_revision_items_invalid_revision_id(self, fxt_get_project, fxt_dataset_service, fxt_client):
+    def test_list_dataset_revision_items_invalid_revision_id(
+        self, fxt_get_project, fxt_dataset_revision_service, fxt_client
+    ):
         response = fxt_client.get(f"/api/projects/{str(fxt_get_project.id)}/dataset_revisions/invalid-id/items")
 
         assert response.status_code == status.HTTP_400_BAD_REQUEST
-        fxt_dataset_service.get_dataset_revision.assert_not_called()
+        fxt_dataset_revision_service.get_dataset_revision.assert_not_called()
 
     def test_get_dataset_revision_item_success(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_item, fxt_dataset_service, fxt_client
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_item, fxt_dataset_revision_service, fxt_client
     ):
-        fxt_dataset_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
+        fxt_dataset_revision_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
         mock_item_data = {
             "id": str(fxt_dataset_item.id),
             "image": f"/path/to/{fxt_dataset_item.name}.jpg",
             "image_info": {"width": fxt_dataset_item.width, "height": fxt_dataset_item.height},
             "subset": fxt_dataset_item.subset.value,
         }
-        fxt_dataset_service.get_dataset_revision_item.return_value = mock_item_data
+        fxt_dataset_revision_service.get_dataset_revision_item.return_value = mock_item_data
 
         response = fxt_client.get(
             f"/api/projects/{str(fxt_get_project.id)}/dataset_revisions/{str(fxt_dataset_revision_id)}/items/{str(fxt_dataset_item.id)}"
@@ -210,18 +218,18 @@ class TestDatasetRevisionItemEndpoints:
         assert response_json["height"] == 768
         assert response_json["subset"] == "training"
 
-        fxt_dataset_service.get_dataset_revision.assert_called_once_with(
+        fxt_dataset_revision_service.get_dataset_revision.assert_called_once_with(
             project_id=fxt_get_project.id, revision_id=fxt_dataset_revision_id
         )
-        fxt_dataset_service.get_dataset_revision_item.assert_called_once_with(
+        fxt_dataset_revision_service.get_dataset_revision_item.assert_called_once_with(
             project_id=fxt_get_project.id, revision_id=fxt_dataset_revision_id, item_id=str(fxt_dataset_item.id)
         )
 
     def test_get_dataset_revision_item_revision_not_found(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_service, fxt_client
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_revision_service, fxt_client
     ):
         dataset_item_id = uuid4()
-        fxt_dataset_service.get_dataset_revision.side_effect = ResourceNotFoundError(
+        fxt_dataset_revision_service.get_dataset_revision.side_effect = ResourceNotFoundError(
             ResourceType.DATASET_REVISION, str(fxt_dataset_revision_id)
         )
 
@@ -230,14 +238,14 @@ class TestDatasetRevisionItemEndpoints:
         )
 
         assert response.status_code == status.HTTP_404_NOT_FOUND
-        fxt_dataset_service.get_dataset_revision_item.assert_not_called()
+        fxt_dataset_revision_service.get_dataset_revision_item.assert_not_called()
 
     def test_get_dataset_revision_item_not_found(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_service, fxt_client
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_revision_service, fxt_client
     ):
         dataset_item_id = uuid4()
-        fxt_dataset_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
-        fxt_dataset_service.get_dataset_revision_item.side_effect = ResourceNotFoundError(
+        fxt_dataset_revision_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
+        fxt_dataset_revision_service.get_dataset_revision_item.side_effect = ResourceNotFoundError(
             ResourceType.DATASET_ITEM, str(dataset_item_id)
         )
 
@@ -248,19 +256,19 @@ class TestDatasetRevisionItemEndpoints:
 
             assert response.status_code == status.HTTP_404_NOT_FOUND
 
-    def test_get_dataset_revision_item_invalid_ids(self, fxt_get_project, fxt_dataset_service, fxt_client):
+    def test_get_dataset_revision_item_invalid_ids(self, fxt_get_project, fxt_dataset_revision_service, fxt_client):
         response = fxt_client.get(
             f"/api/projects/{str(fxt_get_project.id)}/dataset_revisions/invalid-revision-id/items/invalid-item-id"
         )
 
         assert response.status_code == status.HTTP_400_BAD_REQUEST
-        fxt_dataset_service.get_dataset_revision.assert_not_called()
+        fxt_dataset_revision_service.get_dataset_revision.assert_not_called()
 
     def test_get_dataset_revision_item_binary_success(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_service, fxt_client
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_revision_service, fxt_client
     ):
         dataset_item_id = uuid4()
-        fxt_dataset_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
+        fxt_dataset_revision_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
 
         with tempfile.NamedTemporaryFile(suffix=".jpg", delete=False) as tmp_file:
             tmp_file.write(b"test image data")
@@ -268,17 +276,17 @@ class TestDatasetRevisionItemEndpoints:
             tmp_file_path = tmp_file.name
 
         try:
-            fxt_dataset_service.get_dataset_revision_item_binary_path.return_value = tmp_file_path
+            fxt_dataset_revision_service.get_dataset_revision_item_binary_path.return_value = tmp_file_path
 
             response = fxt_client.get(
                 f"/api/projects/{str(fxt_get_project.id)}/dataset_revisions/{str(fxt_dataset_revision_id)}/items/{str(dataset_item_id)}/binary"
             )
 
             assert response.status_code == status.HTTP_200_OK
-            fxt_dataset_service.get_dataset_revision.assert_called_once_with(
+            fxt_dataset_revision_service.get_dataset_revision.assert_called_once_with(
                 project_id=fxt_get_project.id, revision_id=fxt_dataset_revision_id
             )
-            fxt_dataset_service.get_dataset_revision_item_binary_path.assert_called_once_with(
+            fxt_dataset_revision_service.get_dataset_revision_item_binary_path.assert_called_once_with(
                 project_id=fxt_get_project.id, revision_id=fxt_dataset_revision_id, item_id=str(dataset_item_id)
             )
         finally:
@@ -286,10 +294,10 @@ class TestDatasetRevisionItemEndpoints:
                 os.unlink(tmp_file_path)
 
     def test_get_dataset_revision_item_binary_revision_not_found(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_service, fxt_client
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_revision_service, fxt_client
     ):
         dataset_item_id = uuid4()
-        fxt_dataset_service.get_dataset_revision.side_effect = ResourceNotFoundError(
+        fxt_dataset_revision_service.get_dataset_revision.side_effect = ResourceNotFoundError(
             ResourceType.DATASET_REVISION, str(fxt_dataset_revision_id)
         )
 
@@ -298,14 +306,14 @@ class TestDatasetRevisionItemEndpoints:
         )
 
         assert response.status_code == status.HTTP_404_NOT_FOUND
-        fxt_dataset_service.get_dataset_revision_item_binary_path.assert_not_called()
+        fxt_dataset_revision_service.get_dataset_revision_item_binary_path.assert_not_called()
 
     def test_get_dataset_revision_item_binary_not_found(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_service, fxt_client
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_revision_service, fxt_client
     ):
         dataset_item_id = uuid4()
-        fxt_dataset_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
-        fxt_dataset_service.get_dataset_revision_item_binary_path.side_effect = ResourceNotFoundError(
+        fxt_dataset_revision_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
+        fxt_dataset_revision_service.get_dataset_revision_item_binary_path.side_effect = ResourceNotFoundError(
             ResourceType.DATASET_ITEM, str(dataset_item_id)
         )
 
@@ -316,19 +324,21 @@ class TestDatasetRevisionItemEndpoints:
 
             assert response.status_code == status.HTTP_404_NOT_FOUND
 
-    def test_get_dataset_revision_item_binary_invalid_ids(self, fxt_get_project, fxt_dataset_service, fxt_client):
+    def test_get_dataset_revision_item_binary_invalid_ids(
+        self, fxt_get_project, fxt_dataset_revision_service, fxt_client
+    ):
         response = fxt_client.get(
             f"/api/projects/{str(fxt_get_project.id)}/dataset_revisions/invalid-id/items/invalid-item-id/binary"
         )
 
         assert response.status_code == status.HTTP_400_BAD_REQUEST
-        fxt_dataset_service.get_dataset_revision.assert_not_called()
+        fxt_dataset_revision_service.get_dataset_revision.assert_not_called()
 
     def test_get_dataset_revision_item_thumbnail_success(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_service, fxt_client
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_revision_service, fxt_client
     ):
         dataset_item_id = uuid4()
-        fxt_dataset_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
+        fxt_dataset_revision_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
 
         with tempfile.NamedTemporaryFile(suffix=".jpg", delete=False) as tmp_file:
             tmp_file.write(b"test thumbnail data")
@@ -337,17 +347,17 @@ class TestDatasetRevisionItemEndpoints:
 
         try:
             # Thumbnail endpoint now uses the same method as binary (returns full image)
-            fxt_dataset_service.get_dataset_revision_item_binary_path.return_value = tmp_file_path
+            fxt_dataset_revision_service.get_dataset_revision_item_binary_path.return_value = tmp_file_path
 
             response = fxt_client.get(
                 f"/api/projects/{str(fxt_get_project.id)}/dataset_revisions/{str(fxt_dataset_revision_id)}/items/{str(dataset_item_id)}/thumbnail"
             )
 
             assert response.status_code == status.HTTP_200_OK
-            fxt_dataset_service.get_dataset_revision.assert_called_once_with(
+            fxt_dataset_revision_service.get_dataset_revision.assert_called_once_with(
                 project_id=fxt_get_project.id, revision_id=fxt_dataset_revision_id
             )
-            fxt_dataset_service.get_dataset_revision_item_binary_path.assert_called_once_with(
+            fxt_dataset_revision_service.get_dataset_revision_item_binary_path.assert_called_once_with(
                 project_id=fxt_get_project.id, revision_id=fxt_dataset_revision_id, item_id=str(dataset_item_id)
             )
         finally:
@@ -355,11 +365,11 @@ class TestDatasetRevisionItemEndpoints:
                 os.unlink(tmp_file_path)
 
     def test_get_dataset_revision_item_thumbnail_not_found(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_service, fxt_client
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_revision_service, fxt_client
     ):
         dataset_item_id = uuid4()
-        fxt_dataset_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
-        fxt_dataset_service.get_dataset_revision_item_binary_path.side_effect = ResourceNotFoundError(
+        fxt_dataset_revision_service.get_dataset_revision.return_value = MagicMock(id=fxt_dataset_revision_id)
+        fxt_dataset_revision_service.get_dataset_revision_item_binary_path.side_effect = ResourceNotFoundError(
             ResourceType.DATASET_ITEM, str(dataset_item_id)
         )
 
@@ -370,30 +380,32 @@ class TestDatasetRevisionItemEndpoints:
 
             assert response.status_code == status.HTTP_404_NOT_FOUND
 
-    def test_get_dataset_revision_item_thumbnail_invalid_ids(self, fxt_get_project, fxt_dataset_service, fxt_client):
+    def test_get_dataset_revision_item_thumbnail_invalid_ids(
+        self, fxt_get_project, fxt_dataset_revision_service, fxt_client
+    ):
         response = fxt_client.get(
             f"/api/projects/{str(fxt_get_project.id)}/dataset_revisions/invalid-id/items/invalid-item-id/thumbnail"
         )
 
         assert response.status_code == status.HTTP_400_BAD_REQUEST
-        fxt_dataset_service.get_dataset_revision.assert_not_called()
+        fxt_dataset_revision_service.get_dataset_revision.assert_not_called()
 
     def test_delete_dataset_revision_files_success(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_service, fxt_client
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_revision_service, fxt_client
     ):
         response = fxt_client.delete(
             f"/api/projects/{str(fxt_get_project.id)}/dataset_revisions/{str(fxt_dataset_revision_id)}"
         )
 
         assert response.status_code == status.HTTP_204_NO_CONTENT
-        fxt_dataset_service.delete_dataset_revision_files.assert_called_once_with(
+        fxt_dataset_revision_service.delete_dataset_revision_files.assert_called_once_with(
             project_id=fxt_get_project.id, revision_id=fxt_dataset_revision_id
         )
 
     def test_delete_dataset_revision_files_not_found(
-        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_service, fxt_client
+        self, fxt_get_project, fxt_dataset_revision_id, fxt_dataset_revision_service, fxt_client
     ):
-        fxt_dataset_service.delete_dataset_revision_files.side_effect = ResourceNotFoundError(
+        fxt_dataset_revision_service.delete_dataset_revision_files.side_effect = ResourceNotFoundError(
             ResourceType.DATASET_REVISION, str(fxt_dataset_revision_id)
         )
 
@@ -404,8 +416,8 @@ class TestDatasetRevisionItemEndpoints:
 
             assert response.status_code == status.HTTP_404_NOT_FOUND
 
-    def test_delete_dataset_revision_files_invalid_id(self, fxt_get_project, fxt_dataset_service, fxt_client):
+    def test_delete_dataset_revision_files_invalid_id(self, fxt_get_project, fxt_dataset_revision_service, fxt_client):
         response = fxt_client.delete(f"/api/projects/{str(fxt_get_project.id)}/dataset_revisions/invalid-id")
 
         assert response.status_code == status.HTTP_400_BAD_REQUEST
-        fxt_dataset_service.delete_dataset_revision_files.assert_not_called()
+        fxt_dataset_revision_service.delete_dataset_revision_files.assert_not_called()
